Fix linking with libc++ on r12
libc++.{a,so} are not really libraries in r12. They are scripts that
pull in other libraries, such as libunwind, libatomic, and
libandroid_support. The toolchain file was trying to link directly to
c++_{static.a,shared.so}, which no longer works due to symbols having
moved to other libraries (eg. libandroid_support).

This commit fixes the path to the library on r12 and adds the library
path to the list of link paths.